Added dependency injection of IHttpUtil to MLS Refactored the POST method from MLS into HttpUtil Got rid of user-agent and apikey settings in MLS, just pull directly from Prefs
