+++ android/src/main/java/org/mozilla/mozstumbler/service/core/http/HttpUtil.java
@@ -4,6 +4,15 @@
 
 package org.mozilla.mozstumbler.service.core.http;
 
+import android.os.Build;
+
+import org.mozilla.mozstumbler.client.ClientPrefs;
+import org.mozilla.mozstumbler.service.AppGlobals;
+import org.mozilla.mozstumbler.service.core.logging.Log;
+import org.mozilla.mozstumbler.service.utils.Zipper;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -11,18 +20,30 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
 import java.net.Proxy;
 import java.net.ProxySelector;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLConnection;
+import java.util.HashMap;
 import java.util.List;
-
+import java.util.Map;
 
 public class HttpUtil implements IHttpUtil {
 
-    public HttpUtil(){};
+    private static final String LOG_TAG = AppGlobals.LOG_PREFIX + HttpUtil.class.getSimpleName();
+    private static final String USER_AGENT_HEADER = "User-Agent";
+    private final String userAgent;
+
+    public HttpUtil() {
+        this(ClientPrefs.getInstance().getUserAgent());
+    }
+    public HttpUtil(String ua){
+        userAgent = ua;
+    };
 
     private URLConnection openConnectionWithProxy(URL url) throws IOException {
         Proxy proxy = Proxy.NO_PROXY;
@@ -103,5 +124,97 @@ public class HttpUtil implements IHttpUtil {
                 inputStream.close();
             }
         }
+    }
+
+    @Override
+    public IResponse post(String urlString, byte[] data, Map<String, String> headers, boolean precompressed, MLS mls) {
+
+        URL url = null;
+        HttpURLConnection httpURLConnection = null;
+
+        try {
+            url = new URL(urlString);
+        } catch (MalformedURLException e) {
+            throw new IllegalArgumentException("Invalid URL", e);
+        }
+
+        if (data == null) {
+            throw new IllegalArgumentException("Data must be not null");
+        }
+
+        if (headers == null) {
+            headers = new HashMap<String, String>();
+        }
+
+
+        try {
+            httpURLConnection = (HttpURLConnection) url.openConnection();
+            httpURLConnection.setRequestMethod("POST");
+        } catch (IOException e) {
+            Log.e(LOG_TAG, "Couldn't open a connection: " + e);
+            return null;
+        }
+
+        httpURLConnection.setDoOutput(true);
+        httpURLConnection.setRequestProperty(USER_AGENT_HEADER, userAgent);
+        httpURLConnection.setRequestProperty("Content-Type", "application/json");
+
+        // Workaround for a bug in Android mHttpURLConnection. When the library
+        // reuses a stale connection, the connection may fail with an EOFException
+        // http://stackoverflow.com/questions/15411213/android-httpsurlconnection-eofexception/17791819#17791819
+        if (Build.VERSION.SDK_INT > 13 && Build.VERSION.SDK_INT < 19) {
+            httpURLConnection.setRequestProperty("Connection", "Close");
+        }
+
+        // Set headers
+        for (Map.Entry<String, String> entry : headers.entrySet()) {
+            httpURLConnection.setRequestProperty(entry.getKey(), entry.getValue());
+        }
+
+        byte[] wire_data = data;
+        if (!precompressed) {
+            wire_data = Zipper.zipData(data);
+            if (wire_data != null) {
+                httpURLConnection.setRequestProperty("Content-Encoding", "gzip");
+            } else {
+                Log.w(LOG_TAG, "Couldn't compress data, falling back to raw data.");
+                wire_data = data;
+            }
+        } else {
+            httpURLConnection.setRequestProperty("Content-Encoding", "gzip");
+        }
+
+        httpURLConnection.setFixedLengthStreamingMode(wire_data.length);
+        try {
+            OutputStream out = new BufferedOutputStream(httpURLConnection.getOutputStream());
+            out.write(wire_data);
+            out.flush();
+
+            return new HTTPResponse(httpURLConnection.getResponseCode(),
+                    getContentBody(httpURLConnection),
+                    wire_data.length);
+        } catch (IOException e) {
+            Log.e(LOG_TAG, "", e);
+        } finally {
+            if (httpURLConnection != null) {
+                httpURLConnection.disconnect();
+            }
+        }
+        return null;
+    }
+
+    private String getContentBody(HttpURLConnection httpURLConnection) throws IOException {
+        String contentBody;
+        InputStream in = new BufferedInputStream(httpURLConnection.getInputStream());
+        BufferedReader r = new BufferedReader(new InputStreamReader(in));
+        String line;
+        StringBuilder total = new StringBuilder(in.available());
+        while ((line = r.readLine()) != null) {
+            total.append(line);
+        }
+        r.close();
+        in.close();
+        contentBody = total.toString();
+        return contentBody;
     }
 }
