+++ android/src/main/java/org/mozilla/mozstumbler/service/core/http/IHttpUtil.java
@@ -7,7 +7,9 @@ package org.mozilla.mozstumbler.service.core.http;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.Map;
 
 public interface IHttpUtil {
     public String getUrlAsString(URL url) throws IOException;
@@ -17,4 +19,14 @@ public interface IHttpUtil {
     public InputStream getUrlAsStream(String url) throws IOException;
 
     public File getUrlAsFile(URL url, File file) throws IOException;
+
+    /*
+     POST data
+
+    This method will automatically compress any data using gzip
+
+    Return a response object from the server
+    On IOException, this will return null.
+    */
+    IResponse post(String urlString, byte[] data, Map<String, String> headers, boolean precompressed, MLS mls);
 }
